Move tokebabcase to str.ts
